requirements.txt
openai
python-dotenv
fastapi[all]
uvicorn
starlette